delete-one-month-manual.sh: improve the security in the file

+++ delete-one-month-manual.sh
@@ -9,11 +9,24 @@ read year
 echo "Write month: "
 read month
 
-cp $keyword-$year-$month-15* /tmp/
-rm $keyword-$year-$mes* 
+if [ ! -f $keyword-$year-$month-15.sql ]; then
+    echo "$keyword-$year-$month-15.sql not found!. You will not have this backup file"
+fi
 
-cp /tmp/$keyword-$year-$month-15* .
-rm /tmp/$keyword-$year-$month-15* 
+if [ ! -f $keyword-$year-$month-15.tar.gz ]; then
+    echo "$keyword-$year-$month-15.tar.gz not found!. You will not have this backup file"
+fi
 
+wc=`ls -l $keyword-$year-$month* | wc -l`
+echo "There are $wc files to remove. Are you sure?"
+read sure
+
+if [ "$sure" == "yes" ]; 
+then
+    cp $keyword-$year-$month-15* /tmp/
+    rm $keyword-$year-$mes* 
+    cp /tmp/$keyword-$year-$month-15* .
+    rm /tmp/$keyword-$year-$month-15* 
+fi
